Adds deeply nested iframe integration test
