+++ test/msw-api/setup-worker/scenarios/iframe/iframe-deep.page.html
@@ -1,0 +1,1 @@
+<iframe src="./iframe.page.html"></iframe>
