+++ test/msw-api/setup-worker/scenarios/iframe/iframe.page.html
@@ -1,0 +1,1 @@
+<iframe src="./page.html"></iframe>
